Update sign protocol file
